+++ logic/src/main/java/a75f/io/logic/LLights.java
@@ -24,7 +24,7 @@ public class LLights
     {
         for (Output output : zp.getProfileConfiguration(nodeAddress).getOutputs())
         {
-            short dimmablePercent = resolveZoneProfileLogicalValue(zp, output);
+            short dimmablePercent = (short)resolveZoneProfileLogicalValue(zp, output);
             LSmartNode.getSmartNodePort(controlsMessage_t, output.getPort())
                       .set(LSmartNode.mapRawValue(output, dimmablePercent));
         }
